checkpoint: renames "UniformResourceIdentifier/Data/Image/*" to "UniformResourceIdentifier/Image/*"
