+++ src/library/UniformResourceIdentifier/Image/Format/Any.ts
