+++ src/library/UniformResourceIdentifier/Image/Format/all.ts
