+++ src/library/UniformResourceIdentifier/Image/Format/exports.ts
