+++ src/library/UniformResourceIdentifier/Image/Format/index.ts
