+++ src/library/UniformResourceIdentifier/Image/exports.ts
