+++ src/library/UniformResourceIdentifier/Image/index.ts
